[clients] Refactor: renaming datasources to clients
* A more suited term for what it's meant to be use
